--- cgw_push.yaml
@@ -1,122 +1,0 @@
-- type: product
-  state: present
-  metadata:
-    name: "Cloud: Ansible NEW TEST"
-    productCode: "AnsibleNewTest"
-    homepage: "https://test.com/"
-    downloadpage: "https://test.com/"
-    thankYouPage: "https://test.com/"
-    eloquaCode: "NOT_SET"
-    featuredArtifactType: "Server"
-    thankYouTimeout: 5
-- type: product_version
-  state: present
-  metadata:
-    productName: "Cloud: Ansible NEW TEST"
-    productCode: "AnsibleNewTest"
-    versionName: "AnsibleNewTestVersion"
-    ga: true
-    masterProductVersion: null
-    termsAndConditions: "Anonymous Download"
-    trackingDisabled: false
-    hidden: false
-    invisible: false
-    releaseDate: "2022-25-05"
-- type: file
-  state: present
-  metadata:
-    productName: "Cloud: Ansible NEW TEST"
-    productCode: "AnsibleNewTest"
-    productVersionName: "AnsibleNewTestVersion"
-    description: "Red Hat OpenShift Local Sandbox Test"
-    label: "Checksum File Sandbox Test"
-    order: 0
-    hidden: false
-    invisible: false
-    type: "FILE"
-    differentProductThankYouPage: null
-    #    pushItemPath: <repo>/FILES/<filename> # Only one of pushItemPath or downloadURL must present
-    downloadURL: "/content/origin/files/sha256/92/AnsibleNewTest/sha256sum"
-    shortURL: "/pub-1/openshift-v4/clients/AnsibleNewTest/sha256sum"
-    md5: null
-    size: null
-
-- type: product
-  state: absent
-  metadata:
-    name: "Cloud: Ansible TEST New"
-    productCode: "AnsibleTestAgain"
-    homepage: "https://test.com"
-    downloadpage: "https://developers.redhat.com/products/rhsso/download/test/"
-    thankYouPage: "https://developers.redhat.com/products/rhsso/getting-started/test"
-    eloquaCode: "701f2000001D9RsAAK"
-    featuredArtifactType: "Server"
-    thankYouTimeout: 5
-  #
-  #- type: product_version
-  #  state: present
-  #  metadata:
-  #    productName: "Product: JBoss Data Virtualization"
-  #    productCode: "datavirt"
-  #    versionName: "6.0.0 Beta"
-  #    ga: true
-  #    masterProductVersion: null
-  #    termsAndConditions: "JBoss/Red Hat Developer Subscription (obsolete)"
-  #    trackingDisabled: false
-  #    hidden: false
-  #    invisible: false
-  #    releaseDate: "2022-25-05"
-  #    releaseDate: "1470787200000"
-  #- type: file
-  #  state: present
-  #  metadata:
-  #    productName: "Cloud: Ansible TEST New"
-  #    productCode: "AnsibleTestAgain"
-  #    productVersionName: "CGWTest"
-  #    description: "Red Hat OpenShift Local Sandbox Testing"
-  #    label: "Checksum File Sandbox Test"
-  #    order: 0
-  #    hidden: false
-  #    invisible: false
-  #    type: "FILE"
-  #    differentProductThankYouPage: null
-  ##    pushItemPath: <repo>/FILES/<filename> # Only one of pushItemPath or downloadURL must present
-  #    downloadURL: "/content/origin/files/sha256/92/test/sha256sum"
-  #    shortURL: "/pub-1/openshift-v4/clients/ansibletest/CGWTest/sha256sum"
-
-  #- type: file
-  #  state: present|absent
-  #  metadata:
-  #    productName: "Data Virtualization"
-  #    productCode: "datavirt"
-  #    productVersionName: "6.3.0"
-  #    description: "Migration Toolkit CLI"
-  #    label: null
-  #    order: 5
-  #    hidden: true
-  #    invisible: true
-  #    type: "FILE"
-  #    differentProductThankYouPage: 117
-  #    pushItemPath: <repo>/FILES/<filename> # Only one of pushItemPath or downloadURL must present
-  #    downloadURL: "/content/origin/files/sha256/6e/6ee4ffddec425c90b67228c645b00fd61a5d1587c9a6b2ee69af07999b98e265/migrationtoolkit-rhamt-cli-4.0.0.offline.zip"
-  #    shortURL: "/4.0.0/migrationtoolkit-rhamt-cli-4.0.0.offline.zip"
-
-
-  #type: file
-  #state: present
-  #metadata:
-  #  productName: "Data Virtualization"
-  #  productCode: "datavirt"
-  #  productVersionName: "6.3.0"
-  #  description: "Migration Toolkit CLI" #(only for create and update)
-  #  label: null #(only for create and update)
-  #  order: 5 #(only for create and update)
-  #  hidden: false #(only for create and update)
-  #  invisible: false #(only for create and update)
-  #  type: "FILE" #(only for create and update)
-  #  differentProductThankYouPage: 117 #(only for create and update)
-  #  downloadURL: "/content/origin/files/sha256/6e/6ee4ffddec425c90b67228c645b00fd61a5d1587c9a6b2ee69af07999b98e265/migrationtoolkit-rhamt-cli-4.0.0.offline.zip"
-  #  shortURL: "/4.0.0/migrationtoolkit-rhamt-cli-4.0.0.offline.zip" #(only for create and update)
-  #  md5: "31f44cb1302d4603a6ddd3114352414b" #(only for create and update)
-  #  sha256: "6ee4ffddec425c90b67228c645b00fd61a5d1587c9a6b2ee69af07999b98e265" #(only for create and update)
-  #  size: 278064942 #(only for create and update)
